Speed optimization in the case there are no bigfiles in a repo.
Summary:
Now we don't need to talk to S3 in that case.

Use the short-circuit for 'git push' as well.

Test Plan:
Ran 'git bigfile pull' and it was very fast (just long enough for the
'git ls-tree').

Ran 'git bigfile pull' and 'git bigfile push' in webapp and saw it be fast.

Reviewers: chris

Reviewed By: chris

Differential Revision: https://phabricator.khanacademy.org/D17322