If we interrupt a file-download, make sure we don't save the truncated file.
Summary:
I verified that if I did 'git bigfile pull' and then hit control-C in
the middle, it would exit with the current file only partially
downloaded.  We now save the downloaded file in a tempfile,
and atomically rename once the download is successful.

I also make `git bigfile clear` clear out the tempdir where these
temporary files are saved, so they don't just live on forever.

Test Plan:
Ran
       git clone git@github.com:Khan/webapp-i18n-bigfile
       cd webapp-i18n-bigfile/pofiles
       git bigfile pull
       <hit control-C>
and saw that .git/bigfile/tmp had a file in it, and there was
no file in .git/bigfile/objects for the object it said it was
downloading right before the control-C.

I than ran
   git bigfile clear
and saw it say it was deleting 1 item from the tempdir, and
saw there were now no files in .git/bigfile/tmp.

Reviewers: chris, james

Reviewed By: james

Differential Revision: https://phabricator.khanacademy.org/D17327