Speed up many operations by initializing data structures lazily.
GitBigfile.__init__ was calculating all this stuff that was expensive
to compute (often requiring 'git rev-parse' subprocess calls).  And
for many commands, only some of the stuff was needed, or none of it
was.

Now we lazily initialize all that stuff.  This brings down the cost of
a 'git bigfile filter-smudge' from 0.27s to 0.07s.  Still not the
awesome, but we'd probably have to rewrite the code in another
language to do better.

Auditors: chris

Test Plan:
Ran
   git clone git@github.com:Khan/webapp-i18n-bigfile
   cd webapp-i18n-bigfile/
   git bigfile pull
and both commands ran successfully.