vendor boto
Summary:
This will make it so git-bigfile will work no matter what the current
virtualenv is.

Previously we installed boto into the global system python environment,
but when the user was in an isolated virtualenv things could still
break.

After this change is distributed, we will no longer need to install boto
globally as part of khan-dotfiles setup.

(Note the actual vendored directory is going to be amended to the commit
in a second diff to make the actual code changes easier to review.)

#devenvfixup

Test Plan: Verify git-bigfile functions in venv without boto installed.

Reviewers: marcos, csilvers

Reviewed By: csilvers

Subscribers: benkomalo, csilvers, ethan, alpert

Differential Revision: https://phabricator.khanacademy.org/D18534