make symlink aware and package safe
Summary:
use os.path.realpath to make sure we find the path in a symlink aware
way.

additionally, use the same check found in khan-linter to make sure we
don't try to do a git selfupdate when the script is not in an active
git repository (for example, when it has been packaged).

between these two changes it should be safer to version/package this
tool (in homebrew, for example).

Test Plan:
- Make sure git-bigfile still runs properly on a webapp checkout
- Verify homebrew packaged version can run successfully (must be done
post-merge of these changes)

Reviewers: csilvers, ethan

Reviewed By: csilvers, ethan

Projects: #devenvfixup

Differential Revision: https://phabricator.khanacademy.org/D19599