Prevent output of autoupdate git-pull from becoming working copy contents
Summary:
This was manifesting itself as a very confusing error:
https://app.asana.com/0/27216215224639/42514633581592. The contents of a pofile
in intl/translations contained the string "Current branch master is up to date."
instead of the actual file contents.

This turned out to be because the autoupdate mechanism was passing the
auto-update `git pull` output through into the output of the smudge filter!
Craaazy!

The solution is to move the output of the autoupdating to standard error instead
of standard out.

Test Plan:

In the intl/translations directry in webapp:

    rm approved_pofiles/ar.datastore.po
    rm /tmp/git-bigfile.pull
    git checkout -- approved_pofiles/ar.datastore.po
    cat approved_pofiles/ar.datastore.po

Before this change, that would show the following:

    Current branch master is up to date.
    ce5b6380db100fe65ded580299b99d736d45226e

After this change, the "Current branch master is up to date." bit is gone.

Auditors: csilvers, mroth, chelsea